Update README.md
This program is designed to read in an unknown number of ID numbers and test scores from a file. It is then to print the list of students in groups according to letter grade, then print the highest and lowest score along with the corresponding student ID. This program will also compute the average test score of all the tests and print it out.